+++ interfaces/acados_template/acados_template/penalty_utils.py
@@ -1,102 +1,36 @@
+# -*- coding: future_fstrings -*-
+#
+# Copyright (c) The acados authors.
+#
+# This file is part of acados.
+#
+# The 2-Clause BSD License
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# 1. Redistributions of source code must retain the above copyright notice,
+# this list of conditions and the following disclaimer.
+#
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+# this list of conditions and the following disclaimer in the documentation
+# and/or other materials provided with the distribution.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.;
+#
+
 from typing import Optional
-
-from .acados_ocp import AcadosOcp
-from scipy.linalg import block_diag
-
 import casadi as ca
-import numpy as np
-
-
-def formulate_constraint_as_L2_penalty(
-    ocp: AcadosOcp,
-    constr_expr: ca.SX,
-    weight: float,
-    upper_bound: Optional[float],
-    lower_bound: Optional[float],
-    residual_name: str = "new_residual",
-) -> AcadosOcp:
-
-    if upper_bound is None and lower_bound is None:
-        raise ValueError("Either upper or lower bound must be provided.")
-
-    # compute violation expression
-    violation_expr = 0.0
-    if upper_bound is not None:
-        violation_expr = ca.fmax(violation_expr, (constr_expr - upper_bound))
-    if lower_bound is not None:
-        violation_expr = ca.fmax(violation_expr, (lower_bound - constr_expr))
-
-    # add penalty as cost
-    ocp.cost.yref = np.concatenate((ocp.cost.yref, np.zeros(1)))
-    ocp.model.cost_y_expr = ca.vertcat(ocp.model.cost_y_expr, violation_expr)
-    if ocp.cost.cost_type == "NONLINEAR_LS":
-        ocp.cost.W = block_diag(ocp.cost.W, weight)
-    elif ocp.cost.cost_type == "CONVEX_OVER_NONLINEAR":
-        new_residual = ca.SX.sym(residual_name, constr_expr.shape)
-        ocp.model.cost_r_in_psi_expr = ca.vertcat(ocp.model.cost_r_in_psi_expr, new_residual)
-        ocp.model.cost_psi_expr += .5 * weight * new_residual**2
-
-    return ocp
-
-
-def formulate_constraint_as_Huber_penalty(
-    ocp: AcadosOcp,
-    constr_expr: ca.SX,
-    weight: float,
-    upper_bound: Optional[float],
-    lower_bound: Optional[float],
-    residual_name: str = "new_residual",
-    huber_delta: float = 1.0,
-    use_xgn = True,
-) -> AcadosOcp:
-
-    if upper_bound is None and lower_bound is None:
-        raise ValueError("Either upper or lower bound must be provided.")
-    elif upper_bound < lower_bound:
-        raise ValueError("Upper bound must be greater than lower bound.")
-
-    if ocp.cost.cost_type != "CONVEX_OVER_NONLINEAR":
-        raise Exception("Huber penalty is only supported for CONVEX_OVER_NONLINEAR cost type.")
-
-    if (upper_bound is None or lower_bound is None):
-        raise NotImplementedError("only symmetric Huber for now")
-
-    # normalize constraint to [-1, 1]
-    width = upper_bound - lower_bound
-    center = lower_bound + 0.5 * width
-    constr_expr = 2 * (constr_expr - center) / width
-
-    if use_xgn and ocp.model.cost_conl_custom_outer_hess is None:
-        # switch to XGN Hessian start with exact Hessian of previously defined cost
-        exact_cost_hess = ca.hessian(ocp.model.cost_psi_expr, ocp.model.cost_r_in_psi_expr)[0]
-        ocp.model.cost_conl_custom_outer_hess = exact_cost_hess
-
-    # define residual
-    new_residual = ca.SX.sym(residual_name, constr_expr.shape)
-
-    # define penalty
-    penalty, penalty_grad, penalty_hess, penalty_hess_xgn = \
-            symmetric_huber_penalty(new_residual, delta=huber_delta, w=weight*width**2)
-
-    # add penalty to cost
-    ocp.model.cost_r_in_psi_expr = ca.vertcat(ocp.model.cost_r_in_psi_expr, new_residual)
-    ocp.model.cost_psi_expr += penalty
-    ocp.model.cost_y_expr = ca.vertcat(ocp.model.cost_y_expr, constr_expr)
-    ocp.cost.yref = np.concatenate((ocp.cost.yref, np.zeros(1)))
-
-    # add Hessian term
-    if use_xgn:
-        zero_offdiag = ca.SX.zeros(ocp.model.cost_conl_custom_outer_hess.shape[0], penalty_hess_xgn.shape[1])
-        ocp.model.cost_conl_custom_outer_hess = ca.blockcat(ocp.model.cost_conl_custom_outer_hess,
-                                                            zero_offdiag, zero_offdiag.T, penalty_hess_xgn)
-    elif ocp.model.cost_conl_custom_outer_hess is not None:
-        zero_offdiag = ca.SX.zeros(ocp.model.cost_conl_custom_outer_hess.shape[0], penalty_hess_xgn.shape[1])
-        # add penalty Hessian to existing Hessian
-        ocp.model.cost_conl_custom_outer_hess = ca.blockcat(ocp.model.cost_conl_custom_outer_hess,
-                                                            zero_offdiag, zero_offdiag.T, penalty_hess)
-
-    return ocp
-
 
 def huber_loss(var: ca.SX, delta: float, tau: float):
     loss = tau/delta * ca.if_else(
